# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "江都 API",
    "version": "1.1.0",
    "description": "小程序+管理后台共用接口api"
  },
  "host": "localhost:4000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/api/assets/": {
      "post": {
        "tags": [
          "Asset"
        ],
        "summary": "获取asset",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/asset"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Asset"
        ],
        "summary": "修改asset",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/asset_add_update"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Asset"
        ],
        "summary": "删除asset",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/asset_delete"
              }
            }
          }
        }
      }
    },
    "/api/assets/add": {
      "post": {
        "tags": [
          "Asset"
        ],
        "summary": "添加asset",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/asset_add_update"
              }
            }
          }
        }
      }
    },
    "/api/assets/search": {
      "post": {
        "tags": [
          "Asset"
        ],
        "summary": "搜索asset，query参数未知",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ]
      }
    },
    "/api/assets/AssetTags": {
      "post": {
        "tags": [
          "Asset"
        ],
        "summary": "获取标签",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/asset_tag"
              }
            }
          }
        }
      }
    },
    "/api/assets/addTags": {
      "post": {
        "tags": [
          "Asset"
        ],
        "summary": "增加标签",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/asset_tag_add_delete"
              }
            }
          }
        }
      }
    },
    "/api/assets/deleteTags": {
      "post": {
        "tags": [
          "Asset"
        ],
        "summary": "删除标签",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/asset_tag_add_delete"
              }
            }
          }
        }
      }
    },
    "/api/assets/updateStatus": {
      "put": {
        "tags": [
          "Asset"
        ],
        "summary": "？？？？？？？",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/user_register"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "user_register": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "USER_NAME": {
              "type": "string",
              "example": "kevin"
            },
            "PASSWORD": {
              "type": "string",
              "example": "123"
            }
          }
        }
      }
    },
...
    "asset": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {}
        },
        "token": {
          "type": "string",
          "example": "2c59a359-77a0-4b66-9d66-02089946a04c"
        }
      }
    },
    "asset_add_update": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "ASSET_ID": {
              "type": "string",
              "example": "ad52d74e-0ede-11ec-b9ca-342eb7e2a2ec"
            },
            "ASSET_NAME": {
              "type": "string",
              "example": "设备3"
            },
            "ASSET_COUNTRY": {
              "type": "string",
              "example": "日本"
            },
            "MAKER": {
              "type": "string",
              "example": "制造商3"
            },
            "ASSET_COUNT": {
              "type": "number",
              "example": 10
            },
            "START_DATE": {
              "type": "string",
              "example": "2021-09-06 14:50:08.990637000"
            },
            "USER_ID_USER_ID": {
              "type": "string",
              "example": "aa38a206-0a25-11ec-9885-342eb7e2a2ec"
            },
            "USER_NAME": {
              "type": "string",
              "example": "admin"
            },
            "STATUS": {
              "type": "string",
              "example": "已下架"
            },
            "AREA": {
              "type": "string",
              "example": "邗江"
            },
            "FIN_TIME": {
              "type": "string",
              "example": "Invalid date - Invalid date"
            },
            "RATE_RANGE": {
              "type": "string",
              "example": "undefined - undefined"
            },
            "CREATED_AT": {
              "type": "string",
              "example": "Invalid date"
            },
            "UPDATED_AT": {
              "type": "string",
              "example": "Invalid date"
            },
            "vgt_id": {
              "type": "number",
              "example": 0
            },
            "originalIndex": {
              "type": "number",
              "example": 0
            },
            "tags": {
              "type": "array",
              "example": [],
              "items": {}
            }
          }
        },
        "token": {
          "type": "string",
          "example": "2c59a359-77a0-4b66-9d66-02089946a04c"
        }
      }
    },
    "asset_tag": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "ASSET_ID": {
              "type": "string",
              "example": "ad52d74e-0ede-11ec-b9ca-342eb7e2a2ec"
            }
          }
        },
        "token": {
          "type": "string",
          "example": "2c59a359-77a0-4b66-9d66-02089946a04c"
        }
      }
    },
    "asset_delete": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "text": {
              "type": "string",
              "example": "是否要删除undefined(ID: ad52d74d-0ede-11ec-8613-342eb7e2a2ec)?"
            },
            "id": {
              "type": "string",
              "example": "ad52d74d-0ede-11ec-8613-342eb7e2a2ec"
            },
            "type": {
              "type": "string",
              "example": "fin"
            },
            "ASSET_ID": {
              "type": "string",
              "example": "ad52d74d-0ede-11ec-8613-342eb7e2a2ec"
            }
          }
        }
      }
    },
    "asset_tag_add_delete": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "ASSET_ID_ASSET_ID": {
              "type": "string",
              "example": "ad52d74e-0ede-11ec-b9ca-342eb7e2a2ec"
            },
            "TAG_ID_TAG_ID": {
              "type": "string",
              "example": "15cefa4b-0c63-11ec-9ac5-342eb7e2a2ec"
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "JiangduJWT": {
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      }
    }
  }
}